# docs/swagger.yaml
definitions:
  main.List:
    properties:
      address:
        type: string
      birthday:
        type: string
      full_name:
        type: string
      id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Swagger API for Golang Project.
  title: Swagger API
  version: "1.0"
paths:
  /lists:
    get:
      consumes:
      - application/json
      description: Get all
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Show all
      tags:
      - lists
    post:
      consumes:
      - application/json
      description: Add new list
      parameters:
      - description: birthday input format - 1999-01-02
        in: body
        name: newList
        required: true
        schema:
          $ref: '#/definitions/main.List'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: New list
      tags:
      - lists
  /lists/{id}:
    delete:
      consumes:
      - application/json
      description: Delete list
      parameters:
      - description: List ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete by Id
      tags:
      - lists
    put:
      consumes:
      - application/json
      description: Update list
      parameters:
      - description: List ID
        in: path
        name: id
        required: true
        type: string
      - description: List
        in: body
        name: updateList
        required: true
        schema:
          $ref: '#/definitions/main.List'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/main.List'
            type: array
      summary: Update by Id
      tags:
      - lists
swagger: "2.0"